+++ NEWS.md
@@ -1,6 +1,7 @@
 mdsr 0.1.7
 
 * added Connections pane interface to DB Server [#35](https://github.com/beanumber/mdsr/issues/35)
+* removed superfluous extdata [#37](https://github.com/beanumber/mdsr/issues/37)
 
 mdsr 0.1.6